move selector related mongo planner code in separate file
